--Adding PIV Implementation

+++ Code/Run/Misc/mergeCSV.py
@@ -1,44 +1,44 @@
-""" This code merges sweep results across several sweep folders into a sweep folder with all the analysis.csv and .avi files """
 import csv, os, sys, shutil, time
 import numpy as np
+
 def MergeSweeps():
 	path = "..\\..\\Analysis\\"
-	dest = os.path.join(path,'Sweep ' + sys.argv[1])
+	destfolder = 'Sweep ' + sys.argv[1]
 	for folder in os.listdir(path):
-		if 'Sweep' in folder and folder != 'Sweep ' + sys.argv[1]:
-			for dirs, subdirs, files in os.walk(os.path.join(path,folder)):
+		folderpath = os.path.join(path,folder)
+		if 'Sweep' in folder and folder!=destfolder:
+			for dirs, subdirs, files in os.walk(folderpath):
 				for filename in files:
-					filepath = os.path.join(dirs,filename)
-					print filepath
-					sweepindex = filepath.find('Sweep ')
-					endindex = sweepindex + filepath[sweepindex:].find("\\")
-					destpath = filepath[:sweepindex] +'Sweep ' + sys.argv[1] + filepath[endindex:]
-					if os.path.exists(destpath):
-						if 'Analysis' in filename:
+					src_file = os.path.join(dirs,folderpath)
+					dst_dirs = dirs.replace(folder,destfolder)
+					dst_file = os.path.join(dst_dirs, filename)
+					if os.path.exists(dst_file):
+						if "Analysis" in filename:
 							i = 1
-							while os.path.exists(destpath):
-								destpath = destpath[:-4] + str(i) + '.avi'
+							while os.path.exists(dst_file):
+								filename = "Analysis" + str(i) + '.avi'
+								dst_file = os.path.join(dst_dirs,filename)
 								i += 1
-							shutil.copy(filepath,destpath)
-	print 'Done Merging Folders'
-	time.sleep(5)
-
-
+							print "Creating Analysis File:", dst_file
+							shutil.copy(src_file, dst_file)
+					else:
+						if not os.path.exists(dst_dirs):
+							os.makedirs(dst_dirs)
+						print "Creating file:", dst_file
+						shutil.copy(src_file,dst_file)
+	print 'Done Merging Folders\n Commencing CSV Merge'
+	time.sleep(5)			
 
 def Merge(path):
-	csvFiles = []
-	for files in os.listdir(path):
-		if "Analysis" in files:
-			csvFiles.append(files)
-	if len(csvFiles) < 2:
-		return None
-	for csvfile in csvFiles:
-		currFile = os.path.join(path,csvfile)
-		print currFile
-		if os.stat(currFile).st_size == 0:
-			os.remove(currFile)
-		else:
-			with open(os.path.join(path,csvfile),'rb') as f:
+	for csvFile in os.listdir(path):
+		if "Analysis" in csvFile:
+			currFile = os.path.join(path,csvfile)
+			print "Found Analysis File:",currFile
+			if os.stat(currFile).st_size == 0:
+				print "File is empty!"
+				os.remove(currFile)
+				continue
+			with open(currFile,'rb') as f:
 				reader = csv.reader(f)
 				for line in reader:
 					if 'final' not in locals():
@@ -49,18 +49,19 @@ def Merge(path):
 						final = np.concatenate((final,np.array([line])))
 					else:
 						print "Already Exists in Combined Data Set"
-
+	print "Done compiling NumPy Array.\n Sorting and Writing..."
 	if 'final' in locals():
 		final = final[np.lexsort((final[:,7],final[:,6]))]
 		writer = csv.writer(open(os.path.join(path,"Analysis.csv"),'wb'))
 		writer.writerows(final)
 
-	for files in csvFiles:
-		if files != "Analysis.csv":
-			if os.path.exists(os.path.join(path,files)):
+		for files in os.listdir(path):
+			if "Analysis" in files and files != "Analysis.csv":
 				print "Removed:", files
 				os.remove(os.path.join(path,files))
-	return 1
+	else:
+		print "No Data Found!"
+
 
 if __name__ == "__main__":
 	path = "C:\\Users\\clee2\\Documents\\2013 Summer\\Code\\Analysis\\Sweep " + sys.argv[1]
@@ -71,16 +72,17 @@ if __name__ == "__main__":
 	dataset["OneThird"] = ["120","122","124","126","128","130","131","133","134","135","137","139"]
 	dataset["OneQuarter"] = ["141","143","144","145","146","147","148","149","150","151","152"]
 	dataset["OneEighth"] = ["154","155","156","157","159","160","162","163","164","165","166","168","169","170","171","172"]
-	if len(os.listdir("..\\..\\Analysis\\")) > 1:
-		print "Merging Sweep Folders\n"
-		MergeSweeps()
+
+	print "Merging Sweep Folders\n"
+	MergeSweeps()
+	sys.exit()
 
 	for key,values in dataset.items():
 		for value in values:
 			filepath = os.path.join(path,key,value)
 			if not os.path.exists(filepath):
-				print filepath, "does not exist!"
+				continue
 			else:
-				print "Merging Analysis Files\n"
+				print "Merging Analysis Files in %s,%s\n"%(key,value)
 				Merge(filepath)
